front: drop unused props in ManageTimetableItem Map and rename it
