+++ front/src/modules/timetableItem/components/ManageTimetableItem/consts.ts
@@ -1,11 +1,2 @@
-const OPERATIONAL_POINT_LAYERS = [
-  'chartis/osrd_operational_point/geo',
-  'chartis/osrd_operational_point_yardname/geo',
-  'chartis/osrd_operational_point_name_short/geo',
-  'chartis/osrd_operational_point_name/geo',
-  'chartis/osrd_operational_point_kp/geo',
-];
-
-export default OPERATIONAL_POINT_LAYERS;
-
+// eslint-disable-next-line import/prefer-default-export
 export const MAX_TIMEWINDOW_MINUTES = 43800; // 1 month
