+++ front/src/modules/timetableItem/components/ManageTimetableItem/index.ts
@@ -1,3 +1,2 @@
-export { default as Map } from 'modules/timetableItem/components/ManageTimetableItem/Map';
 export { default as TrainSettings } from 'modules/timetableItem/components/ManageTimetableItem/TrainSettings';
 export { default as ElectricalProfiles } from 'modules/timetableItem/components/ManageTimetableItem/ElectricalProfiles';
